// src/types/ChatForUsers.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DtkChat } from "./DtkChat";
import type { DtkChatUser } from "./DtkChatUser";

export interface ChatForUsers { count: number, id: string, chat: Array<DtkChat>, users: Array<DtkChatUser>, }